+++ tdrs-frontend/src/components/SplashPage/SplashPage.test.js
@@ -81,92 +81,6 @@ describe('SplashPage', () => {
     expect(alert.find('h3')).toIncludeText('Inactive Account')
   })
 
-  it('redirects Tribal Coding page when View Tribal Coding Instructions clicked', () => {
-    const store = mockStore(initialState)
-
-    const url =
-      'https://www.acf.hhs.gov/ofa/policy-guidance/tribal-tanf-data-coding-instructions'
-    global.window = Object.create(window)
-    Object.defineProperty(window, 'location', {
-      value: {
-        href: url,
-      },
-    })
-
-    const wrapper = mount(
-      <Provider store={store}>
-        <SplashPage />
-      </Provider>
-    )
-    const link = wrapper.find('#viewTribalCodingInstructions').getElement()
-      .props['href']
-
-    expect(link).toEqual(url)
-  })
-
-  it('redirects to TDP Knowledge Center when View Knowledge Center clicked', async () => {
-    const store = mockStore(initialState)
-
-    const url = 'http://tdp-project-updates.app.cloud.gov/knowledge-center/'
-
-    global.window = Object.create(window)
-    Object.defineProperty(window, 'location', {
-      value: {
-        href: url,
-      },
-    })
-
-    const wrapper = mount(
-      <Provider store={store}>
-        <SplashPage />
-      </Provider>
-    )
-
-    const link = wrapper.find('#viewKnowledgeCenterButton').getElement().props[
-      'href'
-    ]
-
-    expect(link).toEqual(url)
-  })
-
-  it('redirects to ACF Layouts page when View Layouts clicked', () => {
-    const store = mockStore(initialState)
-
-    const url =
-      'https://www.acf.hhs.gov/ofa/policy-guidance/final-tanf-ssp-moe-data-reporting-system-transmission-files-layouts-and-edits'
-    global.window = Object.create(window)
-    Object.defineProperty(window, 'location', {
-      value: {
-        href: url,
-      },
-    })
-
-    const wrapper = mount(
-      <Provider store={store}>
-        <SplashPage />
-      </Provider>
-    )
-    const link = wrapper.find('#viewLayoutsButton').getElement().props['href']
-    expect(link).toEqual(url)
-  })
-
-  it('redirects to ACF Form Instructions when View ACF Form Instructions clicked', () => {
-    const store = mockStore(initialState)
-
-    const url =
-      'https://www.acf.hhs.gov/sites/default/files/documents/ofa/tanf_data_reports_tan_ssp_instructions_definitions.pdf'
-
-    const wrapper = mount(
-      <Provider store={store}>
-        <SplashPage />
-      </Provider>
-    )
-    const link = wrapper.find('#viewACFFormInstructions').getElement().props[
-      'href'
-    ]
-    expect(link).toEqual(url)
-  })
-
   it('redirects to API login endpoint when login.gov sign-in button is clicked', () => {
     const store = mockStore(initialState)
